Renaming Model to Game
